#320: first draft
