+++ package.json
@@ -11,7 +11,6 @@
     "clean": "rm -rf dist",
     "docker-up": "docker compose --env-file development.env up",
     "docker-down": "docker compose --env-file development.env down",
-    "standalone": "node --experimental-network-imports --env-file=development.env dist/jitar.js --config=services/standalone.json --bodylimit=512000",
     "preview": "vite preview",
     "test": "vitest run",
     "test-coverage": "vitest run --coverage",
@@ -19,7 +18,16 @@
     "changelog": "auto-changelog --template changelog.hbs -p -u --commit-limit false --hide-empty-releases true",
     "changelog-debug": "auto-changelog --template changelog.hbs -p --template json --output changelog-data.json",
     "review": "npm run compile && npm run lint && npm run test",
-    "rebuild": "npm run clean && npm run build && npm run standalone"
+    "rebuild": "npm run clean && npm run build && npm run standalone",
+
+    "standalone": "node --experimental-network-imports --env-file=development.env dist/jitar.js --config=services/standalone.json --bodylimit=512000",
+    "repository": "node --experimental-network-imports --env-file=development.env dist/jitar.js --config=services/repository.json --bodylimit=512000",
+    "gateway": "node --experimental-network-imports --env-file=development.env dist/jitar.js --config=services/gateway.json --bodylimit=512000",
+    "bff": "node --experimental-network-imports --env-file=development.env dist/jitar.js --config=services/bff.json --bodylimit=512000",
+    "notification": "node --experimental-network-imports --env-file=development.env dist/jitar.js --config=services/notification.json --bodylimit=512000",
+    "notification2": "node --experimental-network-imports --env-file=development.env dist/jitar.js --config=services/notification2.json --bodylimit=512000",
+    "reads": "node --experimental-network-imports --env-file=development.env dist/jitar.js --config=services/reads.json --bodylimit=512000",
+    "writes": "node --experimental-network-imports --env-file=development.env dist/jitar.js --config=services/writes.json --bodylimit=512000"
   },
   "files": [
     "CHANGELOG.md",
