+++ services/gateway.json
@@ -1,0 +1,17 @@
+{
+    "url": "http://127.0.0.1:3000",
+    "setUp": [
+        "./integrations/runtime/setUpGateway"
+    ],
+    "tearDown": [
+        "./integrations/runtime/tearDownGateway"
+    ],
+    "gateway":
+    {
+        "repository": "http://127.0.0.1:2000",
+        "trustKey": "${JITAR_TRUST_KEY}",
+        "middlewares": [
+            "./integrations/runtime/authenticationMiddleware"
+        ]
+    }
+}
