+++ services/repository.json
@@ -1,0 +1,8 @@
+{
+    "url": "http://127.0.0.1:2000",
+    "repository":
+    {
+        "serveIndexOnNotFound": true,
+        "assets": ["index.html", "main.js", "assets/**/*", "webui/**/*"]
+    }
+}
