+++ services/standalone.json
@@ -18,7 +18,7 @@
         "trustKey": "${JITAR_TRUST_KEY}",
         "serveIndexOnNotFound": true,
         "assets": ["index.html", "main.js", "assets/**/*", "webui/**/*"],
-        "segments": ["webui"],
+        "segments": ["bff", "notification"],
         "middlewares": [
             "./integrations/runtime/authenticationMiddleware"
         ]
