+++ src/domain/creator/create/validateData.ts
