+++ src/domain/creator/register/feature.ts
@@ -1,12 +1,10 @@
 
+import create from '../create/feature';
 import { FULL_NAME_MAX_LENGTH } from '../definitions';
 import generateNickname from '../generateNickname/feature';
 import type { DataModel } from '../types';
 
-import createData from './createData';
 import downloadPortrait from './downloadPortrait';
-import insertData from './insertData';
-import validateData from './validateData';
 
 export default async function feature(fullName: string, nickname: string, email: string, portraitUrl?: string): Promise<DataModel>
 {
@@ -17,11 +15,5 @@ export default async function feature(fullName: string, nickname: string, email:
         ? await downloadPortrait(portraitUrl)
         : undefined;
 
-    const data = await createData(truncatedFullName, generatedNickname, email, portraitId);
-
-    validateData({ fullName, email, portraitId });
-
-    await insertData(data);
-
-    return data;
+    return create(truncatedFullName, generatedNickname, email, portraitId);
 }
